Fix : Error In Friends Serailezer

+++ services/player/api/views.py
@@ -6,7 +6,7 @@ from django.utils.decorators import method_decorator
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
-from .serializers import PlayerSerializer, PlayerInfoSerializer
+from .serializers import PlayerInfoSerializer
 from .models import Player, Friendship
 from .decorators import jwt_cookie_required
 import urllib.parse
@@ -25,7 +25,7 @@ class PlayerInfo(APIView):
                 return Response({
                     "status": 200,
                     "players": serializer.data,
-                    "message": "User found successfully",
+                    "message": "User found successfully"
                 })
             player = Player.objects.get(id=request.decoded_token['id'])
             serializer = PlayerInfoSerializer(player)
@@ -140,8 +140,15 @@ class PlayerFriendship(APIView):
                     for friendship in friendships:
                         friend = friendship.sender
                         friend_data = {
+                            "id": friend.id,
+                            "first_name": friend.first_name,
+                            "last_name": friend.last_name,
                             "username": friend.username,
+                            "status": friend.status,
+                            "tournament_name": friend.tournament_name,
+                            "two_factor": friend.two_factor,
                             "avatar": friend.avatar
+                            
                         }
                         friendship_data.append(friend_data)
                     return Response({
@@ -149,36 +156,35 @@ class PlayerFriendship(APIView):
                         "friendships": friendship_data
                     })
                 elif get_type == 'friends':
-                    friendships = Friendship.objects.filter(sender=id, status='AC')
+                    friendships = Friendship.objects.filter(status='AC')
                     friendship_data = []
+                    encountered_usernames = [] 
                     for friendship in friendships:
-                        friend = friendship.receiver
-                        friend_data = {
-                            "username": friend.username,
-                            "avatar": friend.avatar,
-                            "id": friend.id,
-                            "status": friend.status,
-                            "tournament_name": friend.tournament_name,
-                            "two_factor": friend.two_factor,
-                            "email": friend.email,
-                            "first_name": friend.first_name,
-                            "last_name": friend.last_name
-                        }
-                        friendship_data.append(friend_data)
-                        friend = friendship.sender
-                        friend_data = {
-                            "username": friend.username,
-                            "avatar": friend.avatar,
-                            "id": friend.id,
-                            "status": friend.status,
-                            "tournament_name": friend.tournament_name,
-                            "two_factor": friend.two_factor,
-                            "email": friend.email,
-                            "first_name": friend.first_name,
-                            "last_name": friend.last_name
-                        }
-                        if friend_data not in friendship_data:
+                        sender_username = friendship.sender.username
+                        receiver_username = friendship.receiver.username
+                        
+                        if sender_username not in encountered_usernames:
+                            friend_data = {
+                                "username": sender_username,
+                                "avatar": friendship.sender.avatar
+                            }
                             friendship_data.append(friend_data)
+                            encountered_usernames.append(sender_username)
+                        
+                        if receiver_username not in encountered_usernames:
+                            friend_data = {
+                                "id": friendship.receiver.id,
+                                "username": receiver_username,
+                                "first_name": friendship.receiver.first_name,
+                                "last_name": friendship.receiver.last_name,
+                                "avatar": friendship.receiver.avatar,
+                                "status": friendship.recevier.status,
+                                "tounament_name": friendship.receiver.tournament_name,
+                                "two_factor": friendship.receiver.two_factor
+                            }
+                            friendship_data.append(friend_data)
+                            encountered_usernames.append(receiver_username)
+
                     return Response({
                         "status": 200,
                         "friendships": friendship_data
@@ -189,31 +195,16 @@ class PlayerFriendship(APIView):
                     for friendship in friendships:
                         friend = friendship.receiver
                         friend_data = {
-                            "username": friend.username,
-                            "avatar": friend.avatar,
                             "id": friend.id,
+                            "username": friend.username,
+                            "first_name": friend.first_name,
+                            "last_name": friend.last_name,
                             "status": friend.status,
                             "tournament_name": friend.tournament_name,
-                            "two_factor": friend.two_factor,
-                            "email": friend.email,
-                            "first_name": friend.first_name,
-                            "last_name": friend.last_name
+                            "avatar": friend.avatar,
+                            "two_factor": friend.two_factor
                         }
                         friendship_data.append(friend_data)
-                        friend = friendship.sender
-                        friend_data = {
-                            "username": friend.username,
-                            "avatar": friend.avatar,
-                            "id": friend.id,
-                            "status": friend.status,
-                            "tournament_name": friend.tournament_name,
-                            "two_factor": friend.two_factor,
-                            "email": friend.email,
-                            "first_name": friend.first_name,
-                            "last_name": friend.last_name
-                        }
-                        if friend_data not in friendship_data:
-                            friendship_data.append(friend_data)
                     return Response({
                         "status": 200,
                         "friendships": friendship_data
@@ -274,24 +265,26 @@ class PlayerFriendship(APIView):
 
         @method_decorator(jwt_cookie_required)
         def delete(self, request):
-            try :
-                id = request.decoded_token['id']
-                sender = Player.objects.get(id=id)
-                receiver_id = request.data.get('target_id')
-                receiver = Player.objects.get(id=receiver_id)
-                friendship = Friendship.objects.get(sender=sender, receiver=receiver)
-                friendship.delete()
-                return Response({
-                    "status": 200,
-                    "message": 'Friendship deleted successfully'
-                })
-            except Friendship.DoesNotExist:
-                return Response({
+            id = request.decoded_token['id']
+            sender = Player.objects.get(id=id)
+            receiver_id = request.data.get('target_id')
+            receiver = Player.objects.get(id=receiver_id)
+            friendship = Friendship.objects.get(sender=sender, receiver=receiver)
+            if friendship is None:
+                try :
+                    friendship = Friendship.objects.get(sender=receiver, receiver=sender)
+                    friendship.delete()
+                    return Response({
+                        "status": 200,
+                        "message": 'Friendship deleted successfully'
+                    })
+                except Friendship.DoesNotExist:
+                    return Response({
                     "status": 404,
                     "message": "Friend request not found",
                 })
-            except Exception as e:
-                return Response({
+                except Exception as e:
+                    return Response({
                     "status": 500,
                     "message": str(e),
                 })
